krampol/requirements.txt
disnake
python-dotenv
requests
asyncio